Added Schema classes as abstraction
