+++ .gitignore
@@ -85,6 +85,10 @@ ipython_config.py
 # pyenv
 .python-version
 
+# pylint / pyreverse
+*.dot
+*.png
+
 # pipenv
 #   According to pypa/pipenv#598, it is recommended to include Pipfile.lock in version control.
 #   However, in case of collaboration, if having platform-specific dependencies or dependencies
